Move query editing commands to separate file
